ast_analysis: Add `new_behavior` flag for detecting different result.
Execute `analysis_iter_anchoring` twice, with a flag unset and set,
and compare the result. The `env->past_any_consuming` field is only
updated when the flag is set.

Just checking for the flag's behavior within the `AST_EXPR_ALT` case
isn't enough, because it will have lots of false positives. We need to
check whether the different result there caused the overall AST to
be flagged as unsatisfiable. The bug is that libfsm should have been
able to tell `^.^` is unsatisfiable (can never match anything), but
it was instead treating it like `.`, a regex that matches anything.