re: Add flags for generating input, capture resolution, multi-regexes.
This was pretty much the minimal amount I needed for manual testing.

- `-FC`: No captures, do not build capture metadata even when dialects
  support it.

- Generating input: `build/bin/re -rpcre -G10 '^a*b*$'`

- Capture resolution: `build/bin/re -rpcre -R '^a(b*)c$' abbbbbc`

    -- 0: 0,7
    -- 1: 1,6

- Multiple regexes: `build/bin/re -rpcre -pgC -Y file_with_one_regex_per_line`

Capture resolution is not implemented for -y or multi-regex yet, because
`fsm_exec_with_captures` needs all input buffered ahead of time and
because multi-regex isn't passing along the capture bases and matching
those up based on the result.