Disable the fragile IR tests.
These are fragile due to testing the graph structure when rendered to json literally. The problem here is the state numbering and order matter, and that depends on the compiler's code generation, malloc, current wind direction, etc.

What we should be doing here is comparion of the graph's structure, but I don't know how to do that.

We're effectively covering the same situations through end-to-end tests for the generated code, so I'm just going to disable these until I come up with a better way to deal with the IR structure.