ir: Wrap ids and count in a struct, to namespace them as endids.
The IR struct is about to get another id & count pair.

This loses storing the count as a 31-bit bitfield, but if the
goal for that is saving memory then the ids array allocation could
be replaced with a struct that contains the count, and then each
IR without endids will save more space than the current approach.