cdata: Remove assertions on realloc's result.
I confirmed that the callers actually check the return.

These should probably use the alloc interface (with `f_realloc`),
but the existing print callback typedefs don't seem to pass along
the alloc handle anymore, since it was removed from fsm_options,
so if that changes it will be in a different commit.