tmp: Add EXPENSIVE_CHECKS wrapper around fsm_exec's is_isdfa check.
Normally this doesn't matter, because the expected API use is based on
code generation rather than heavily using `fsm_exec`, but we are
calling fsm_exec to check that the combined DFAs still match all of
their original regexes. This is very expensive. (It similarly gets
very cumbersome during fuzzing.)

A better way to address this long-term is tracking on the FSM whether
it's result of fsm_determinise (and optionally fsm_minimise), and
updating that info to indicate that it reverted to an NFA if any
other operations modify the FSM states/edges. That's outside the scope
of this PR though.