cdata: Address warning, ensure closing `}` is always output.
Previously `AMBIG_NONE` was missing the closing curly brace but
the other variants added it. All three need it, so it's cleaner
to move it outside the switch case.

Also, make AMBIG_NONE output `(void)endid_base;`, because otherwise it
leads to a warning about the unused variable. (It's probably not worth
complicating the output template further to avoid using it in the
first place here.)