Experiment: Make 0 implicit in the bitset_words table.
A word of 0 is by far the most common (since most regexes will only
use printable ASCII), so making it implicit removes one slot from
the head of the bitset_words entries that are likely to stay in cache.
Instead, use any ID >= the bitset_words table count to represent 0.
This doubles as a range check.